Add some fixes
