+++ .github/workflows/upgrade-cli.yml
@@ -88,14 +88,14 @@ jobs:
             \"PS_DISABLE_OVERRIDES\":\"true\"
           }' > modules/autoupgrade/config.json"
 
-      - name: Update
+      - name: Update online
         if: matrix.UPGRADE_CHANNEL == 'online'
         run: |
           docker exec -t prestashop sh -c "echo 'memory_limit=1024M' > /usr/local/etc/php/conf.d/zz-memory-limit.ini"
           docker exec -t prestashop php modules/autoupgrade/bin/console update:start --channel=${{ matrix.UPGRADE_CHANNEL }} admin-dev
           docker exec -t prestashop chmod 777 -R /var/www/html/var
 
-      - name: Update
+      - name: Update local
         if: matrix.UPGRADE_CHANNEL == 'local'
         run: |
           docker exec -t prestashop sh -c "echo 'memory_limit=1024M' > /usr/local/etc/php/conf.d/zz-memory-limit.ini"
