+++ tests/UI/package.json
@@ -16,11 +16,11 @@
         "// Sanity": "Scripts to run sanity tests",
         "test:sanity": "npx playwright test /sanity --workers 1 --max-failures 1",
         "// upgradeOnline": "Scripts to run functional tests to upgrade online channel",
-        "test:upgradeOnline": "npx playwright test /functional/01_upgradeOnline.spec.ts --workers 1 --max-failures 1",
+        "test:upgradeOnline": "npx playwright test /functional/01_checkModal.spec.ts /functional/02_upgradeOnline.spec.ts --workers 1 --max-failures 1",
         "// upgradeLocal": "Scripts to run functional tests to upgrade local channel",
-        "test:upgradeLocal": "npx playwright test /functional/02_upgradeLocal.spec.ts --workers 1 --max-failures 1",
+        "test:upgradeLocal": "npx playwright test /functional/01_checkModal.spec.ts /functional/03_upgradeLocal.spec.ts --workers 1 --max-failures 1",
         "// rollback": "Scripts to run functional tests to rollabck",
-        "test:rollback": "npx playwright test /functional/03_rollback.spec.ts --workers 1 --max-failures 1"
+        "test:rollback": "npx playwright test /functional/04_rollback.spec.ts --workers 1 --max-failures 1"
     },
     "keywords": [],
     "devDependencies": {
